refactor: 💡 Ethers removed from exports; Types for ethers-v6
