+++ .eslintignore
@@ -11,3 +11,5 @@ cache-zk
 artifacts
 cache
 ./**/*.sol
+typechain
+typechainV6
