+++ .lintstagedrc
@@ -1,4 +1,4 @@
 {
-  "*.{ts,js}": ["prettier --write", "eslint --cache --fix"],
+  "*.{ts,js}": ["prettier --write", "eslint --config ./.eslintrc --ignore-path ./.eslintignore --cache --fix"],
   "*.md": ["prettier --write"]
 }
