+++ .prettierignore
@@ -13,3 +13,4 @@ artifacts
 cache
 network
 typechain
+typechainV6
