+++ package.json
@@ -89,6 +89,7 @@
     "typechain": "^8.1.1",
     "@typechain/hardhat": "^6.1.6",
     "@typechain/ethers-v5": "^10.2.1",
+    "@typechain/ethers-v6": "^0.3.2",
     "dotenv": "^16.0.3"
   },
   "dependencies": {
@@ -96,7 +97,7 @@
   },
   "scripts": {
     "build:contracts": "hardhat compile",
-    "build:src": "rm -rf ./lib && tsc -p ./tsconfig.build.json",
+    "build:src": "rm -rf ./lib && typechain --target=ethers-v6 --out-dir typechainV6 artifacts/**/Config.json artifacts/**/EntitiesRegistry.json artifacts/**/Market.json && tsc -p ./tsconfig.build.json",
     "test:contracts": "hardhat test",
     "test:src": "mocha -t 60000 --extension spec.ts test/src",
     "lint": "solhint . && eslint --ext .ts",
